Prepare APIView for Testing
