+++ src/dotnet/APIView/APIViewWeb/Controllers/AutoReviewController.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using APIViewWeb.Filters;
-using APIViewWeb.Repositories;
+using APIViewWeb.Managers;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
