+++ src/dotnet/APIView/APIViewWeb/Controllers/ReviewController.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using APIViewWeb.Filters;
-using APIViewWeb.Repositories;
+using APIViewWeb.Managers;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.ApplicationInsights.Extensibility;
