+++ src/dotnet/APIView/APIViewWeb/Controllers/UserProfileController.cs
@@ -3,6 +3,7 @@ using APIViewWeb.Models;
 using APIViewWeb.Repositories;
 using System.Threading.Tasks;
 using System.Collections.Generic;
+using APIViewWeb.Managers;
 
 namespace APIViewWeb.Controllers
 {
@@ -32,7 +33,7 @@ namespace APIViewWeb.Controllers
         [HttpPost]
         public async Task<ActionResult> Update(string email, string[] languages, string theme="light-theme")
         {
-            UserProfileModel profile = await _userProfileManager.tryGetUserProfileAsync(User);
+            UserProfileModel profile = await _userProfileManager.TryGetUserProfileAsync(User);
             UserPreferenceModel preference = await _userPreferenceCache.GetUserPreferences(User);
 
             preference.Theme = theme;
@@ -42,10 +43,10 @@ namespace APIViewWeb.Controllers
 
             if(profile.UserName == null)
             {
-                await _userProfileManager.createUserProfileAsync(User, email, Languages, preference);
+                await _userProfileManager.CreateUserProfileAsync(User, email, Languages, preference);
             } else
             {
-                await _userProfileManager.updateUserProfile(User, email, Languages, preference);
+                await _userProfileManager.UpdateUserProfile(User, email, Languages, preference);
             }
             this._userPreferenceCache.UpdateUserPreference(preference, User);
