+++ src/dotnet/APIView/APIViewWeb/HostedServices/PullRequestBackgroundHostedService.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using APIViewWeb.Repositories;
+using APIViewWeb.Managers;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.Extensions.Configuration;
