+++ src/dotnet/APIView/APIViewWeb/Managers/IPackageNameManager.cs
@@ -1,0 +1,11 @@
+using APIViewWeb.Models;
+using System.Threading.Tasks;
+
+namespace APIViewWeb.Managers
+{
+    public interface IPackageNameManager
+    {
+        public Task<PackageModel> GetPackageDetails(string packageName);
+
+    }
+}
