+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Conversation.cshtml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using ApiView;
 using APIView;
+using APIViewWeb.Managers;
 using APIViewWeb.Models;
 using APIViewWeb.Repositories;
 using Microsoft.AspNetCore.Mvc;
