+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Delete.cshtml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using APIViewWeb.Managers;
 using APIViewWeb.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
