+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/Samples.cshtml.cs
@@ -5,6 +5,7 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using ApiView;
 using APIView;
+using APIViewWeb.Managers;
 using APIViewWeb.Models;
 using APIViewWeb.Repositories;
 using Microsoft.AspNetCore.Authorization;
