+++ src/dotnet/APIView/APIViewWeb/Repositories/BlobCodeFileRepository.cs
@@ -6,20 +6,21 @@ using System.IO;
 using System.Threading.Tasks;
 using ApiView;
 using APIViewWeb.Models;
+using APIViewWeb.Repositories;
 using Azure.Storage.Blobs;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Configuration;
 
 namespace APIViewWeb
 {
-    public class BlobCodeFileRepository
+    public class BlobCodeFileRepository : IBlobCodeFileRepository
     {
         private BlobContainerClient _container;
         private readonly IMemoryCache _cache;
 
-        public BlobCodeFileRepository(IConfiguration configuration, IMemoryCache cache, BlobContainerClient blobContainerClient = null)
+        public BlobCodeFileRepository(IConfiguration configuration, IMemoryCache cache)
         {
-            _container = blobContainerClient ?? new BlobContainerClient(configuration["Blob:ConnectionString"], "codefiles");
+            _container = new BlobContainerClient(configuration["Blob:ConnectionString"], "codefiles");
             _cache = cache;
         }
 
@@ -51,19 +52,6 @@ namespace APIViewWeb
             return codeFile;
         }
 
-        private BlobClient GetBlobClient(string revisionId, string codeFileId, out string key)
-        {
-            if (revisionId == codeFileId)
-            {
-                key = codeFileId;
-            }
-            else
-            {
-                key = revisionId + "/" + codeFileId;
-            }
-            return _container.GetBlobClient(key);
-        }
-
         public async Task UpsertCodeFileAsync(string revisionId, string codeFileId, CodeFile codeFile)
         {
             var memoryStream = new MemoryStream();
@@ -82,6 +70,19 @@ namespace APIViewWeb
         {
             await GetBlobClient(revisionId, codeFileId, out var key).DeleteAsync();
             _cache.Remove(key);
+        }
+
+        private BlobClient GetBlobClient(string revisionId, string codeFileId, out string key)
+        {
+            if (revisionId == codeFileId)
+            {
+                key = codeFileId;
+            }
+            else
+            {
+                key = revisionId + "/" + codeFileId;
+            }
+            return _container.GetBlobClient(key);
         }
     }
 }
