+++ src/dotnet/APIView/APIViewWeb/Repositories/BlobOriginalsRepository.cs
@@ -4,31 +4,27 @@
 using System;
 using System.IO;
 using System.Threading.Tasks;
+using APIViewWeb.Repositories;
 using Azure.Storage.Blobs;
 using Microsoft.Extensions.Configuration;
 
 namespace APIViewWeb
 {
-    public class BlobOriginalsRepository
+    public class BlobOriginalsRepository : IBlobOriginalsRepository
     {
         private BlobContainerClient _container;
 
         public string GetContainerUrl() => _container.Uri.ToString();
 
-        public BlobOriginalsRepository(IConfiguration configuration, BlobContainerClient blobContainerClient = null)
+        public BlobOriginalsRepository(IConfiguration configuration)
         {
-            _container = blobContainerClient ?? new BlobContainerClient(configuration["Blob:ConnectionString"], "originals");
+            _container = new BlobContainerClient(configuration["Blob:ConnectionString"], "originals");
         }
 
         public async Task<Stream> GetOriginalAsync(string codeFileId)
         {
             var info = await GetBlobClient(codeFileId).DownloadAsync();
             return info.Value.Content;
-        }
-
-        private BlobClient GetBlobClient(string codeFileId)
-        {
-            return _container.GetBlobClient(codeFileId);
         }
 
         public async Task UploadOriginalAsync(string codeFileId, Stream stream)
@@ -39,6 +35,11 @@ namespace APIViewWeb
         public async Task DeleteOriginalAsync(string codeFileId)
         {
             await GetBlobClient(codeFileId).DeleteAsync();
+        }
+
+        private BlobClient GetBlobClient(string codeFileId)
+        {
+            return _container.GetBlobClient(codeFileId);
         }
     }
 }
