+++ src/dotnet/APIView/APIViewWeb/Repositories/CosmosUserProfileRepository.cs
@@ -1,30 +1,27 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using APIViewWeb.Models;
-using Microsoft.AspNetCore.Mvc;
+using APIViewWeb.Repositories;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Serialization.HybridRow;
 using Microsoft.Extensions.Configuration;
-using Microsoft.VisualStudio.Services.Graph.Client;
 
 namespace APIViewWeb
 {
-    public class CosmosUserProfileRepository
+    public class CosmosUserProfileRepository : ICosmosUserProfileRepository
     {
         private readonly Container _userProfileContainer;
 
-        public CosmosUserProfileRepository(IConfiguration configuration, CosmosClient cosmosClient = null)
+        public CosmosUserProfileRepository(IConfiguration configuration)
         {
-            var client = cosmosClient ?? new CosmosClient(configuration["Cosmos:ConnectionString"]);
+            var client = new CosmosClient(configuration["Cosmos:ConnectionString"]);
             _userProfileContainer = client.GetContainer("APIView", "Profiles");
         }
 
-        public async Task<UserProfileModel> tryGetUserProfileAsync(string UserName)
+        public async Task<UserProfileModel> TryGetUserProfileAsync(string UserName)
         {
             try
             {
@@ -36,7 +33,7 @@ namespace APIViewWeb
             }
         }
 
-        public async Task<Result> upsertUserProfileAsync(ClaimsPrincipal User, UserProfileModel userModel)
+        public async Task<Result> UpsertUserProfileAsync(ClaimsPrincipal User, UserProfileModel userModel)
         {
             if(User.GetGitHubLogin().Equals(userModel.UserName))
             {
