+++ src/dotnet/APIView/APIViewWeb/Repositories/ICosmosPullRequestsRepository.cs
@@ -1,0 +1,14 @@
+using APIViewWeb.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace APIViewWeb.Repositories
+{
+    public interface ICosmosPullRequestsRepository
+    {
+        public Task<PullRequestModel> GetPullRequestAsync(int pullRequestNumber, string repoName, string packageName, string language = null);
+        public Task UpsertPullRequestAsync(PullRequestModel pullRequestModel);
+        public Task<IEnumerable<PullRequestModel>> GetPullRequestsAsync(bool isOpen);
+        public Task<List<PullRequestModel>> GetPullRequestsAsync(int pullRequestNumber, string repoName);
+    }
+}
